+++ src/idvc/utils/AutomaticRegistration.py
